UBUNTU: SAUCE: iperf: ensure interval for mpstat is an int
Going from python2 to python3 resulted in a float being returned when
calculation the mpstat interval. Force this value to int for mpstat.

Acked-by: Cory Todd <cory.todd@canonical.com>
Acked-by: Sean Feole <sean.feale@canonical.com>
Acked-by: Po-Hsu Lin <po-hsu.lin@canonical.com>

+++ iperf/iperf.py
@@ -125,7 +125,7 @@ class iperf(test.test):
             # Get 5 mpstat samples. Since tests with large number of streams
             # take a long time to start up all the streams, we'll toss out the
             # first and last sample when recording results
-            interval = max(1, test_time / 5)
+            interval = int(max(1, test_time / 5))
             cmds.append('mpstat -P ALL %s 5' % interval)
 
             # Add the iperf command
